requirements_mac.txt
absl-py==2.1.0
altgraph==0.17.4
asttokens==2.4.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
colour==0.1.5
contourpy==1.2.0
cryptography==42.0.1
cycler==0.12.1
decorator==5.1.1
Deprecated==1.2.14
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.1
fonttools==4.47.2
fsspec==2023.12.2
gcloud==0.18.3
google-auth==2.27.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
httplib2==0.22.0
idna==3.6
importlib-metadata==7.0.1
ipython==8.20.0
jaraco.classes==3.3.0
jedi==0.19.1
Jinja2==3.1.3
jwcrypto==1.5.1
jwt==1.3.1
keyring==24.3.0
kiwisolver==1.4.5
macholib==1.16.3
Markdown==3.5.2
MarkupSafe==2.1.4
matplotlib==3.8.2
matplotlib-inline==0.1.6
more-itertools==10.2.0
MouseInfo==0.1.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.23.5
oauth2client==4.1.3
oauthlib==3.2.2
opencv-python==4.9.0.80
packaging==23.2
pandas==2.2.0
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
prompt-toolkit==3.0.43
protobuf==4.21.2
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pycparser==2.21
pycryptodome==3.20.0
PyGetWindow==0.0.9
Pygments==2.17.2
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2024.0
PyMsgBox==1.0.9
pyobjc-core==10.1
pyobjc-framework-Cocoa==10.1
pyobjc-framework-Quartz==10.1
pyparsing==3.1.1
pyperclip==1.8.2
Pyrebase4==4.7.1
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.8.2
python-jwt==4.1.0
pytweening==1.0.7
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.29.0
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rsa==4.9
rubicon-objc==0.4.7
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
stack-data==0.6.3
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tkmacosx==1.0.5
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
tqdm==4.66.1
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==1.26.18
wcwidth==0.2.13
Werkzeug==3.0.1
wrapt==1.16.0
zipp==3.17.0
